requirements.txt
ibm-cloud-sdk-core
ibm_watson
python-vlc
